+++ lib/saytime_weather/weather_weatherapi.rb
@@ -159,10 +159,5 @@ module SaytimeWeather
       end
     end
 
-    def mph_to_ms(mph)
-      return nil unless mph && mph.is_a?(Numeric)
-
-      mph / 2.23694
-    end
   end
 end