+++ lib/saytime_weather/weather_wttr.rb
@@ -70,11 +70,6 @@ module SaytimeWeather
       }
     end
 
-    def mph_to_ms(mph)
-      return nil unless mph && mph.is_a?(Numeric)
-      mph / 2.23694
-    end
-
     def wttr_text_to_condition(text)
       SaytimeWeather::WeatherConditions.from_text(text)
     end